Client timeout option added; MIDI port/channel display (fix).
git-svn-id: http://svn.code.sf.net/p/qsampler/code_attic/trunk@82 3e92087d-0f25-4bc2-b788-d05a45ecc5c1

+++ src/qsamplerOptions.cpp
@@ -141,7 +141,7 @@ qsamplerOptions::~qsamplerOptions (void)
 
     // Recent file list.
     m_settings.beginGroup("/RecentFiles");
-    for (int i = 0; i < recentFiles.count(); i++)
+    for (int i = 0; i < (int) recentFiles.count(); i++)
         m_settings.writeEntry("/File" + QString::number(i + 1), recentFiles[i]);
     m_settings.endGroup();
